refactor: Add user info loading from JSON file
This commit refactors the `bot.py` file to include loading user information from a JSON file. It imports the `json` module and the `UserInfo` class from `userinfo.py`. The `userinfo` dictionary is populated with data from the `/data/userinfo.json` file. Additionally, for each member in the guild, if their ID is not present in the `userinfo` dictionary, a new entry is created with a default value for the `messageCount` key.